avoid compiler warning (cast enum to int)

+++ core/vil/file_formats/vil_geotiff_header.cxx
@@ -164,7 +164,7 @@ bool vil_geotiff_header::get_key_value(geokey_t key, void** value,
   // check if it is a valid key id
   if (length == 0) {
     // key is not defined
-    vcl_cerr << "KeyID=" << key << " is not valid\n";
+    vcl_cerr << "KeyID=" << (short int)key << " is not valid\n";
     return false;
   }
   else {
